[*] refactoring code
